fix comments by javier
